+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Renamed OPEN-SCRIPT to SCRIPTS
 - Moved DELETE-SCRIPT functionality to `Ctrl`+`X` under SCRIPTS
 - Added `y` key binding to copy in LINKS
+- Changed to using `<` and `>` for rotation of key binding info grid
+- Added bottom right pane toggle via `?`
 - Made the UX of various input fields slightly more polished
     - This includes adding `Esc` for cancelling during SAVE-SCRIPT
 - Added request handling debouncing to reduce pointless workload triggered during